requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.4
idna==3.4
importlib-resources==5.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
packaging==21.3
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.3
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.6
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.9.0
